Improvements to xas_spectra
nexus_scan.py
 - added hdf_find and hdf_tree functions
 - xas_scan -> xas_spectra

spectra_container.py
 - refactored nexus writer methods into nexus_reader.py
 - removed decorator and directly implemented spectra process methods

nxxas_loader.py
 - added dls_loader to load_xas_scans and related functions.
